Claude Code project setup; pin ruff rules, fix dangling state-emit task
CLAUDE.md, an add-provider skill, a conventions skill, and two subagents
(realtime-audit, test-writer), plus a PostToolUse ruff hook. The hook runs
lint fixes only: this codebase is hand-formatted and diverges from ruff's
formatter in 21 files, and F401 autofix deletes imports a multi-step edit
has added but not yet used.

pyproject pinned an explicit lint select. Ruff's implicit default widens
between releases, so with only >=0.8 and no select the lint surface moved
under us — a fresh install reported 51 errors. B008 and RUF001 are ignored
with reasons: FastAPI's DI idiom is a call in an argument default, and the
Devanagari danda and Arabic full stop are sentence terminators we segment
on, not confusable typos.

Fixes a real defect found by the resulting lint pass: _set_state spawned
the state emit with no reference held. The loop only weakly references a
running task, so a state change could be collected before it was sent and
the console would miss it intermittently. Tasks are now owned until they
complete, and _finish waits briefly for the last one before cancelling.

Also: zip(strict=True) in the vector store, where _cosine over mismatched
embedding dimensions silently truncated to the shorter vector instead of
failing; audioop-lts declared for 3.13, which requires-python allows but
where the import would otherwise fail; and the test fixture moved off
idle timeouts the Settings bounds reject, which had left 13 tests erroring.

+++ vaani/audio/resample.py
@@ -27,7 +27,7 @@ def resample_pcm16(pcm: bytes, src_rate: int, dst_rate: int) -> bytes:
     samples = np.frombuffer(pcm, dtype=np.int16)
     if samples.size == 0:
         return b""
-    out_len = int(round(samples.size * dst_rate / src_rate))
+    out_len = round(samples.size * dst_rate / src_rate)
     if out_len <= 0:
         return b""
     src_idx = np.linspace(0, samples.size - 1, num=out_len, dtype=np.float64)
